+++ examples-pydantic-v1/omni_multi_pydantic_v1/_api_client.py
@@ -111,7 +111,7 @@ class OmniMultiClient:
     OmniMultiClient
 
     Generated with:
-        pygen = 0.99.32
+        pygen = 0.99.33
         cognite-sdk = 7.54.13
         pydantic = 1.10.7
 
@@ -125,7 +125,7 @@ class OmniMultiClient:
         else:
             raise ValueError(f"Expected CogniteClient or ClientConfig, got {type(config_or_client)}")
         # The client name is used for aggregated logging of Pygen Usage
-        client.config.client_name = "CognitePygen:0.99.32"
+        client.config.client_name = "CognitePygen:0.99.33"
 
         self.omni_multi_a = OmniMultiAAPIs(client)
         self.omni_multi_b = OmniMultiBAPIs(client)