+++ examples-pydantic-v1/omni_sub_pydantic_v1/_api_client.py
@@ -23,7 +23,7 @@ class OmniSubClient:
     OmniSubClient
 
     Generated with:
-        pygen = 0.99.32
+        pygen = 0.99.33
         cognite-sdk = 7.54.13
         pydantic = 1.10.7
 
@@ -41,7 +41,7 @@ class OmniSubClient:
         else:
             raise ValueError(f"Expected CogniteClient or ClientConfig, got {type(config_or_client)}")
         # The client name is used for aggregated logging of Pygen Usage
-        client.config.client_name = "CognitePygen:0.99.32"
+        client.config.client_name = "CognitePygen:0.99.33"
 
         self._client = client