+++ tests/test_unit/test_models/test_fields.py
@@ -62,6 +62,7 @@ def test_field_from_property_expect_warning(name: str, expected_name, pygen_conf
             dm.ViewId("a", "b", "c"),
             pydantic_field="Field",
             has_default_instance_space=True,
+            direct_relations_by_view_id={},
         )
 
     # Assert